cleaned up index.html, some typos in angular-leaflet
Main error was a missing jquery.min.map file. I downloaded
jquery-2.0.3.min.map and renamed the file to match the missing
reference. Removed a GoogleMaps API call that's not needed.

Currently in global settings, i have this setup to call for leaflet
maps, b/c I don't have the is4cwn tiles. To run it locally on the app
server, we should swap that selection.